+++ tests/HomeScreen.test.tsx
@@ -1,5 +1,8 @@
 import { TestUserPosts } from "@lib/posts";
-import HomeScreen, { HomeScreenViewReply } from "@screens/HomeScreen";
+import HomeScreen, {
+  homeScreenViewedReplyFromRouteParams,
+  HomeScreenViewReply,
+} from "@screens/HomeScreen";
 import { fireEvent, render, screen } from "@testing-library/react-native";
 import "./helpers/Matchers";
 import { Button, View } from "react-native";
@@ -46,3 +49,38 @@ const renderHomeScreen = (viewedReply?: HomeScreenViewReply) => {
 const dismissReplyScreen = () => fireEvent.press(screen.getByText("dismiss"));
 
 const replyScreen = () => screen.queryByTestId("viewed-reply");
+
+const testPostId = "post";
+const testReplyId = "reply";
+
+describe("homeScreenViewedReplyFromRoute tests", () => {
+  it("should return a HomeViewedReply when both a post and reply id are set", () => {
+    const viewedReply = homeScreenViewedReplyFromRouteParams({
+      postId: testPostId,
+      replyId: testReplyId,
+    });
+    expect(viewedReply).toMatchObject({
+      postId: testPostId,
+      replyId: testReplyId,
+    });
+  });
+
+  it("should return undefined when undefined params", () => {
+    const viewedReply = homeScreenViewedReplyFromRouteParams(undefined);
+    expect(viewedReply).toBeUndefined();
+  });
+
+  it("should return undefined when only post id", () => {
+    const viewedReply = homeScreenViewedReplyFromRouteParams({
+      postId: testPostId,
+    });
+    expect(viewedReply).toBeUndefined();
+  });
+
+  it("should return undefined when only reply id", () => {
+    const viewedReply = homeScreenViewedReplyFromRouteParams({
+      replyId: testReplyId,
+    });
+    expect(viewedReply).toBeUndefined();
+  });
+});
